Fixed gravship substructure bugs
NOTE: This does change behavior, removing the behavior where rich soil, stony soil, etc. generate with regular soil underneath, and causes placing soil on top of other soil to delete the soil underneath. However, placing soil on top of soil used to delete soil in certain circumstances anyways due to the way this patch is written, and I'm unsure that rich/stony soil was intended to generate with regular soil underneath in the first place.

Tested the fix with only DLC/Harmony/Consistent Map Stone active and was able to build substructure and land a gravship with no problem on rich/stony soil.

Note that this requires a new map, previously generated maps already have the underTerrain generated and thus will not work.